requirements-dev.txt
black==20.8b1
coverage==5.4
flake8==3.8.4
pytest==6.2.2
pytest-cov==2.11.1
requests-mock==1.8.0
